refactor: organizes DescriptionList exports in their respective folder
