+++ src/components/ui/description-list/index.ts
@@ -1,1 +1,3 @@
 export { default as DescriptionList } from './DescriptionList.vue';
+export { default as DescriptionListItem } from './DescriptionListItem.vue';
+export type { DescriptionListItemProps } from './types';
