+++ src/components/ui/description-list/types.ts
@@ -1,0 +1,4 @@
+export interface DescriptionListItemProps {
+  label: string;
+  description?: string;
+}
